requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
databases==0.6.1
ecdsa==0.18.0
exceptiongroup==1.0.4
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.1
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
install==1.3.5
MarkupSafe==2.1.1
ormar==0.12.0
packaging==21.3
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
Werkzeug==2.2.2
